Move Goal and DerivationGoal out of build.cc
Make the build more incremental and the file easier to digest
